Introduce `UserSystem`
