+++ app3d/src/jsMain/kotlin/diy/lingerie/web_tool_3d/UserSystem.kt
@@ -1,0 +1,5 @@
+package diy.lingerie.web_tool_3d
+
+data class UserSystem(
+    val userBezierMesh: UserBezierMesh,
+)
